refactor(utils): Rename metrics/get-metric-type to utils/metrics
