--- packages/utils/src/metrics/index.js
@@ -1,1 +1,0 @@
-export * from './get-metric-type';
