+++ packages/utils/src/utils/index.js
@@ -1,3 +1,4 @@
 export * from './delta';
 export * from './date';
 export * from './format';
+export * from './metrics';
