+++ packages/utils/src/utils/metrics.js
@@ -2,6 +2,20 @@ import { get } from 'lodash';
 
 import METRICS, { METRIC_TYPE_NUMERIC, METRIC_TYPES } from '../config/metrics';
 
+/**
+ * Get metric type data
+ *
+ * @param {String} key Metric key
+ * @param {String} [type] Default metric type
+ *
+ * @typedef {Object} Metric
+ * @property {String} label Metric label
+ * @property {String} type Metric type
+ * @property {Function} formatter Metric format handler
+ * @property {Boolean} biggerIsBetter Metric flag
+ *
+ * @return @Metric
+ */
 export const getMetricType = (key, type) => {
   const metric = get(METRICS, key);
